requirements.txt
# Core Dependencies
jinja2>=3.1.2
pyyaml>=6.0.1
weasyprint>=60.1
pandas>=2.0.0
streamlit>=1.28.0

# Advanced PDF & Document Processing
reportlab>=4.0.7
python-docx>=0.8.11
pillow>=10.0.0
pypdf>=3.17.0

# Security & Encryption
cryptography>=41.0.0
python-dotenv>=1.0.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Image Processing & EXIF Handling
piexif>=1.1.3
exifread>=3.0.0

# Data Validation
pydantic>=2.0.0
email-validator>=2.0.0

# Utilities & Helpers
python-slugify>=8.0.0
arrow>=1.3.0
requests>=2.31.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Development & Testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Optional: For async processing
aiofiles>=23.2.1
asyncio-contextmanager>=1.0.0
